server/requirements.txt
flask
redis
requests
spacy
dict.cc.py
textblob
py-openthesaurus
wiktionaryparser